update view,fix permission

+++ tesapp/views.py
@@ -17,44 +17,71 @@ class CartContentViewSet(viewsets.ModelViewSet):
 
     @action(detail=True, methods=['get'])
     def get_by_CartId(self, request, pk=None):
-        queryset = CartContent.objects.filter(cartId=pk)
+        user = request.user
+        queryset = CartContent.objects.filter(cart=pk)
+        cart = Cart.objects.get(id=pk)
+        if cart.user.pk != user.pk: 
+            return Response('You are not authorized to access this cart.', status=403)
+        temp = []
+        # for i in  queryset:
+        #     temp = [i]
         serializer = CartContentSerializer(queryset, many=True)
         return Response(serializer.data)
-    
     @action(detail=False, methods=['post'])
     def get_by_MenuId_CartId(self, request):
         data = request.data
+        user = request.user
         menuId = data['menuId']
         cartId = data['cartId']
-
         cartContent = CartContent.objects.get(cart=cartId, menu=menuId)
-
+        if cartContent.cart.user.pk != user.pk:
+            return Response('You are not authorized to access this cart.', status=403)
+        
         serializer = CartContentSerializer(cartContent)
         return Response(serializer.data)
-        
+
     @action(detail=False, methods=['post'])
-    def update_quantity_by_cartId_menuId(self, request):
+    def add_quantity_by_cartId_menuId(self, request):
         data = request.data
+        user = request.user
         cartId = data['cartId']
         menuId = data['menuId']
         newQuantity = data['quantity']
 
-        cartContent = CartContent.objects.filter(cartId__exact=cartId).get(menuId__exact=menuId)
-        cartContent.quantity = newQuantity
+        cartContent = CartContent.objects.get(cart=cartId,menu=menuId)
+        if cartContent.cart.user.pk != user.pk:
+            return Response('You are not authorized to access this cart.', status=403)
+        
+        cartContent.quantity = cartContent.quantity + newQuantity
         cartContent.save()
 
         serializer = CartContentSerializer(data=model_to_dict(cartContent))
         if serializer.is_valid():
             return Response(serializer.data)
         return Response(serializer.errors)
-        
+    @action(detail=True,methods=['post'])
+    def update_quantity(self,request,pk):
+        data = request.data
+        user = request.user
+        cartContent = CartContent.objects.get(id=pk)
+        if cartContent.cart.user.pk != user.pk:
+            return Response('You are not authorized to update this cart.', status=403)
+        cartContent.quantity = data['quantity']
+        cartContent.save()
+        serializer = CartContentSerializer(data=model_to_dict(cartContent))
+        if serializer.is_valid():
+            return Response(serializer.data)
+        return Response(serializer.errors)
     @action(detail=False, methods=['post'])
     def delete_by_CartId_MenuId(self, request):
         data = request.data
+        user = request.user
         cartId = data['cartId']
         menuId = data['menuId']
-
-        cartContent = CartContent.objects.filter(cartId__exact=cartId).filter(menuId__exact=menuId)
+        
+        cartContent = CartContent.objects.get(cart=cartId,menu=menuId)
+        if cartContent.cart.user.pk != user.pk:
+            return Response('You are not authorized to delete this cart.', status=403)
         cartContent.delete()
         return Response("berhasil dihapus")
 
@@ -63,10 +90,22 @@ class CartViewSet(viewsets.ModelViewSet):
     serializer_class = CartSerializer
     http_method_class = ['get', 'post', 'put', 'delete']
     permission_classes = (IsCustomer, )
-
+    @action(detail=False,methods=['get'])
+    def add(self,request):
+        user = request.user
+        cart = Cart(user=user)
+        cart.save()
+        serializer = CartSerializer(data=model_to_dict(cart))
+        if serializer.is_valid():
+            return Response(serializer.data)
+        return Response(serializer.errors)
     @action(detail=True, methods=['get'])
     def set_checkout_true_by_id(self, request, pk):
-        cart = Cart.objects.get(id__exact=pk)
+        cart = Cart.objects.get(id=pk)
+        user = request.user
+        if cart.user.pk != user.pk:
+            return Response('You are not authorized to update this cart.', status=403)
+        
         cart.checkedOut = True
         cart.checkOutTime = timezone.now()
         cart.save()
@@ -79,10 +118,14 @@ class CartViewSet(viewsets.ModelViewSet):
     @action(detail=False, methods=['post'])
     def update_status_by_id(self, request):
         data = request.data
+        user = request.user
         cartId = data['id']
         newStatus = data['status']
 
-        cart = Cart.objects.get(id__exact=cartId)
+        cart = Cart.objects.get(id=cartId)
+        if cart.user.pk != user.pk:
+            return Response('You are not authorized to update this cart.', status=403)
+        
         cart.status = newStatus
         cart.save()
 
@@ -94,6 +137,8 @@ class CartViewSet(viewsets.ModelViewSet):
     @action(detail=True,methods=['get'])
     def delete(self, request, pk):
         cart = Cart.objects.get(id__exact=pk)
+        if cart.user.pk != user.pk:
+            return Response('You are not authorized to delete this cart.', status=403)
         cart.delete()
         return Response("cart berhasil dihapus")
